chore: cleanup test name
